+++ parallel-consumer-core/pom.xml
@@ -129,7 +129,7 @@
                         <package>io.confluent.csid</package>
                     </packages>
                     <classes>
-                        <class>io.confluent.csid.actors.Actor</class>
+                        <class>io.confluent.csid.actors.ActorImpl</class>
                         <class>io.confluent.parallelconsumer.PollContext</class>
                         <class>io.confluent.parallelconsumer.ParallelEoSStreamProcessor</class>
                         <class>io.confluent.parallelconsumer.internal.ProducerManager</class>
