+++ parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/internal/BrokerPollSystem.java
@@ -4,7 +4,7 @@ package io.confluent.parallelconsumer.internal;
  * Copyright (C) 2020-2022 Confluent, Inc.
  */
 
-import io.confluent.csid.actors.Actor;
+import io.confluent.csid.actors.ActorImpl;
 import io.confluent.parallelconsumer.ParallelConsumerOptions;
 import io.confluent.parallelconsumer.state.WorkManager;
 import lombok.*;
@@ -69,7 +69,7 @@ public class BrokerPollSystem<K, V> implements OffsetCommitter {
     private final WorkManager<K, V> wm;
 
     @Getter
-    private final Actor<BrokerPollSystem<K, V>> myActor = new Actor<>(this);
+    private final ActorImpl<BrokerPollSystem<K, V>> myActor = new ActorImpl<>(this);
 
     public BrokerPollSystem(ConsumerManager<K, V> consumerMgr, WorkManager<K, V> wm, ControllerInternalAPI<K, V> pc, final ParallelConsumerOptions<K, V> options) {
         this.wm = wm;
